# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gaming Lounge OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products/menu items
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /product-categories:
    get:
      operationId: listProductCategories
      tags: [products]
      summary: List product categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductCategory"
    post:
      operationId: createProductCategory
      tags: [products]
      summary: Create a product category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductCategory"

  # ── QR MENU ──────────────────────────────────────────
  /qr/{token}/menu:
    get:
      operationId: getQrMenu
      tags: [qr]
      summary: Get menu by QR token (public, no auth)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Menu with asset info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrMenuResponse"
        "404":
          description: Token not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /qr/{token}/order:
    post:
      operationId: placeQrOrder
      tags: [qr]
      summary: Place an order via QR (public, no auth)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QrOrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── ORDERS ──────────────────────────────────────────
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, verified]
        - name: sessionId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a payment record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{paymentId}/verify:
    post:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify a payment (cashier confirms)
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentVerifyInput"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ── SHIFTS ──────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Asset:
      type: object
      required: [id, name, type, pricePerHour, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: ["string", "null"]
        type:
          type: string
          enum: [ps, billiard, air_hockey, babyfoot, other]
        pricePerHour:
          type: number
        status:
          type: string
          enum: [available, busy]
        qrToken:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Session:
      type: object
      required: [id, assetId, status, startedAt]
      properties:
        id:
          type: integer
        assetId:
          type: integer
        assetName:
          type: ["string", "null"]
        assetNameAr:
          type: ["string", "null"]
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, paused, ended, cancelled]
        startedAt:
          type: string
          format: date-time
        pausedAt:
          type: ["string", "null"]
          format: date-time
        endedAt:
          type: ["string", "null"]
          format: date-time
        totalMinutes:
          type: ["number", "null"]
        totalCost:
          type: ["number", "null"]
        cancelReason:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    Product:
      type: object
      required: [id, name, price, isAvailable, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        categoryNameAr:
          type: ["string", "null"]
        price:
          type: number
        isAvailable:
          type: boolean
        description:
          type: ["string", "null"]
        descriptionAr:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ProductInput:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
        nameAr:
          type: string
        categoryId:
          type: integer
        price:
          type: number
        description:
          type: string
        descriptionAr:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        categoryId:
          type: integer
        price:
          type: number
        description:
          type: string
        descriptionAr:
          type: string
        isAvailable:
          type: boolean
    ProductCategory:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ProductCategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        nameAr:
          type: string

    # ── QR MENU SCHEMAS ──
    QrMenuResponse:
      type: object
      required: [asset, categories, products]
      properties:
        asset:
          $ref: "#/components/schemas/Asset"
        categories:
          type: array
          items:
            $ref: "#/components/schemas/ProductCategory"
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        activeSession:
          $ref: "#/components/schemas/Session"
    QrOrderInput:
      type: object
      required: [items]
      properties:
        customerName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        notes:
          type: string

    # ── ORDER SCHEMAS ──
    OrderItem:
      type: object
      required: [id, productId, productName, quantity, unitPrice, totalPrice]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productNameAr:
          type: ["string", "null"]
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
        notes:
          type: ["string", "null"]
    Order:
      type: object
      required: [id, source, status, totalAmount, items, createdAt]
      properties:
        id:
          type: integer
        source:
          type: string
          enum: [qr, pos]
        status:
          type: string
          enum: [pending, preparing, ready, delivered, closed, cancelled]
        sessionId:
          type: ["integer", "null"]
        assetId:
          type: ["integer", "null"]
        assetName:
          type: ["string", "null"]
        assetNameAr:
          type: ["string", "null"]
        createdByUserId:
          type: ["integer", "null"]
        createdByUserName:
          type: ["string", "null"]
        assignedToUserId:
          type: ["integer", "null"]
        assignedToUserName:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        totalAmount:
          type: number
        cancelReason:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
        preparingAt:
          type: ["string", "null"]
          format: date-time
        readyAt:
          type: ["string", "null"]
          format: date-time
        deliveredAt:
          type: ["string", "null"]
          format: date-time
...
    Payment:
      type: object
      required: [id, sessionId, method, amount, status, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        method:
          type: string
          enum: [cash, instapay, visa]
        amount:
          type: number
        status:
          type: string
          enum: [pending, verified]
        transactionReference:
          type: ["string", "null"]
        instapayReference:
          type: ["string", "null"]
        verifiedByUserId:
          type: ["integer", "null"]
        verifiedByUserName:
          type: ["string", "null"]
        verifiedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    PaymentInput:
      type: object
      required: [sessionId, method, amount]
      properties:
        sessionId:
          type: integer
        method:
          type: string
          enum: [cash, instapay, visa]
        amount:
          type: number
        transactionReference:
          type: string
    PaymentVerifyInput:
      type: object
      properties:
        transactionReference:
          type: string

    # ── SHIFT SCHEMAS ──
...