# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gaming Lounge OS API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  nameAr?: string;
  categoryId?: number;
  price: number;
  description?: string;
  descriptionAr?: string;
}